\name{AmeliaView}
\alias{AmeliaView}
\title{Interactive GUI for Amelia}
\description{Brings up the AmeliaView graphical interface, which allows users
  to load datasets, manage options and run Amelia from a traditional 
  windowed environment.
}
\usage{
AmeliaView
}
\keyword{ utilities }% at least one, from doc/KEYWORDS
