\name{freetrade}
\docType{data}
\alias{freetrade}

\title{Trade Policy and Democracy in 9 Asian States}

\description{Economic and political data on nine developing countries in
Asia from 1980 to 1999. This dataset includes 9 variables including
year, country, average tariff rates, Polity IV score, total population,
gross domestic product per capita, gross international reserves, a dummy
variable for if the country had signed an IMF agreement in that year, a
measure of financial openness, and a measure of US hegemony. These data
were used in Milner and Kubota (2005).}

\usage{freetrade}
\format{A data frame with 10 variables and 171 observations.}
\source{World Bank, World Trade Organization, Polity IV and others.}
\references{Helen Milner and Keiko Kubota (2005), ``Why the move to free trade?
  Democracy and trade policy in the developing countries.''
  \emph{International Organization}, Vol 59, Issue 1.}
\keyword{datasets}
