\name{summary.amelia}
\alias{summary.amelia}

\title{Summary of an Amelia object}

\description{Returns summary information from the Amelia run along with
  missingles information.}

\usage{
\method{summary}{amelia}(object, ...)
}
\arguments{
  \item{object}{an object of class \code{amelia}. Typically, an output
    from the function \code{amelia}.}
  \item{...}{further arguments.}
}

\seealso{\code{\link{amelia}}, \code{\link{plot.amelia}}}
